[DomCrawler] Abstract URI-related logic
All the URI parsing logic is externalized in the AbstractUriElement
class, implementing the UriElementInterface interface.

The AbstractUriElement class have two abstract methods:

* setNode: validate the DOMElement node according to the concrete class
  rules, and set $this->node.
* getRawUri: get the raw URI from $this->node.

The Link and Form classs now extends AbstractUriElement (Form don't have
to extend Link anymore).

This refactor is desirable for #12429.
